Build a new requirements list
This is the result of running a script (see next commit) which just
takes all of the pip-requires and test-requires files from our
projects and makes a consolidated list from them.

I had to do some manual cleanup to remove duplicates. This means it's
not a complete list since some of the requires are actually incompatible
like keystoneclient or sqlalchemy.

So, this still isn't sufficient to build a complete PyPI mirror from
yet, but it is closer to the current requirements set and doesn't
artificially pin us to specific versions.

We need to come back around and cap the version ranges according to
each upstream's policy for changing their release numbers when they
make incompatible changes:

  http://lists.openstack.org/pipermail/openstack-dev/2013-March/006331.html

Change-Id: I616aa67eee6c559906b2a7a2322686e0c6d846da