pin pep8 to an exact version
prior to an aggreed pinning of pep8 in grizzly we ended up with a
rediculous amount of code churn on minor pep8 bumps accross the
projects. It was seen as a huge step forward to pin this to a version
for a release.

We need to not regress this and reintroduce the code churn on pep8
bumps happening upstream, as every single one of them typically
requires a code change on our behalf to integrate. We should take
a single hit at the beginning of a release cycle to the newest pep8
version, and keep it pinned at that till the next release.

Change-Id: I957e98f7e82156a26c86ee5d55c6993486070404