Add sphinxcontrib-pecanwsme
Ceilometer had a Sphinx extension in our
code tree that we use to auto-generate
the API documentation from the Pecan/WSME
code. That extension is being moved to
a separate package so other projects can
use it more easily, and so we now need
the package added to the requirements list.

The change to use the package in ceilometer is
https://review.openstack.org/29148.

Change-Id: I07ad3a12c7420feeb6296e1b6ad982a910bbc03d
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>