Add requirement for PBR level >=0.5.10,<0.6
There is a bug in PBR that I discovered while doing Cinder development
on Havana.  It causes the migrate.cfg file to not be properly
setup in /usr/lib/python2.X/site-packages/cinder/db/sqlalchemy/migrate_repo
which results in the database not being properly populated.  That was
seen with pbr==0.5.8.

This change is proposing making a more specific requirement of pbr>=0.5.10 as
the problem encountered with 0.5.8 is resolved with 0.5.10 and later
versions of PBR.

Change-Id: I4ee527331f65751621d0763910d48f963697bbd2
fixes: bug 1182203