Cap requests <1.2.1.
* tools/pip-requires: The requests library released 1.2.1 today
which brings in sphinx as a new dependency. One of sphinx's
transitive dependencies is on a newer version of jinja than is
provided in the python-jinja2 package on CentOS 6, and since nova's
unit tests allow site-packages in the venv for libvirt support this
cap is needed so that gate-nova-python26 will continue to work. A
corresponding patch is being drafted for nova, bug 1182271.

Change-Id: I2ea71d2bbc8a4e67a4c07672aadef218330641da