Add lower bound for all dependencies with upper bound
If we care about a max version of a package, also
specify the min version to have a range to check for.
Specify min version always in a '>=' fashion.
This simplifies a bit of the handling, as the first
given range is always the lower bound.

Also raise websockify to 0.5.1 in preparation of
https://review.openstack.org/#/c/35365/

Change-Id: Ia2672d675446f2ce9932c0c8b09e7b3c2c9cc392