Allow use of oslo.config-1.2.0a3
See bug #1194807

We add a new dev-requirements.txt file because we don't want anything
not released to pypi to show up with the openstack pypi mirror which
is built from requirements.txt and test-requirements.txt.

Neutron and oslo-incubator already uses oslo.config 1.2.0a2. Nova used
it briefly but we hit an issue.

1.2.0a3 contains a fix for a regression in the registration of the oslo
namespace package, so it doesn't make sense to allow a2.

Change-Id: I6f3eb5fd2c75615d9a1cae172aed859b36b27d4c