remove python-fooclient version caps
per mailing list thread, allowing any project to cap one of our
own clients actually prevents us from testing that client's git
version in devstack gate (because there are 2 ways to install a
python client, one from git, one from pip). This can allow bugs
to creep into the clients because they are no longer tested in
the gate.

We need to start with uncapping the global list, then ensuring
we uncap this in all the projects.

Neutron client excluded because of a need for incompatible
changes in v3.

Needed to make progress on bug #1200214

Change-Id: Ice57c9bd9ce2bbd786ca0c80a1b94049a7f63ff8