Remove version pins from setup_requires
setuptools can never upgrade things based on setup_requires versions.
What this means is that pbr can NEVER EVER introduce a feature or a
change that would require specifying a newer version of pbr in the
setup.py file, because if will break. This is a non-flexible
requirement.

The requirements.txt file can catch the actual version we're interested
in having around, if it's an install time requirement.

Change-Id: Ic491f295156ea0ee3f6296360ca8b31d5134ae07