Add ordereddict for python 2.6 compatibility
Glance property protections (bp/api-v2-property-protection) is
predicated on sections in a configuration file being read in the order
they appear.

This is the default behaviour for ConfigParser in Python 2.7, however
for Python 2.6 we need to specify the 'dict_type' parameter
(ref
http://docs.python.org/2/library/configparser.html#ConfigParser.RawConfigParser
)

This dependency can be removed when OpenStack is no longer required to
support Python 2.6.

Change-Id: Iafe9d1ffe2062156c5194b491c7fe55b130a7461