add an explanation to the sqla requirement
people keep trying to "fix" this to a broken statement, and I now
realize the reasons for this being added are basically lore lost
in the mists of time. So document it.

The wrong fix is no longer blocked in our gate because we force
install newer pip on our devstack nodes (otherwise lots of other
things break).

Change-Id: I048e33d101a0c09adabac6d17cbc0ff23fba19e7