Set a floor on the python-memcache library
python-memcache prior to version 1.48 will leak memory due to always
caching cas ids (compare and set) and with nothing designed to clean
it up.  In 1.48 it was changed to require a boolean argument 'cache_cas'
on init to enable the cas id caching.  In the case of keystone, we rely
on the cache_cas argument and cas functionality.  Ideally, all uses of
memcache should use later than 1.47 to prevent the memory leaking of
cas id caching.

Change-Id: Ide780e30a8243cd5fcc693515ffd84dfea3a3467