Exclude iso8601 version 0.1.7
The iso8601 package had a 0.1.7 release on October 19th.
Unfortunately, it cannot be installed under Python 3.3 due to a
unicode character in its README.rst, which is read by its
setup.py.  This causes our py33 gates to fail.  The temporary
solution is to exclude 0.1.7.  As a more permanent solution, I
have reported this as a bug[1] to the iso8601 developer, which
should hopefully result in a fix.

[1] https://bitbucket.org/micktwomey/pyiso8601/issue/10/setuppy-broken-for-python-33

Change-Id: Icdd8223b26caf755c8d6cd4c27c99ecea5f54a75