Adopt 0.1.8 as iso8601 minimum version
In the images api, we depend on iso8601 to parse some dates and times.
Recently, since version 0.1.4, python-iso8601 added support for a few
more formats, and we finally got some other issues nailed down by 0.1.8.
If end user switch from 0.1.8 back to 0.1.4 in the deployment, the image
api will lose support for certain datetime formats like YYYY-MM-DD
(where the time part is assumed to be all zeros). This obviously creates
a compatibility concern.

We were resolved to make glance compatible with both iso 0.1.4 and
0.1.8, but it appears that to do so requires essentially reimplementing
a large swath of iso8601 0.1.8 in glance itself.

Fixes bug: 1242501

Change-Id: Iece0116d3c412216251e2d7fbffe10bfe2c1c789
Signed-off-by: Zhi Yan Liu <zhiyanl@cn.ibm.com>