Add decorator to the global requirements
This module is usefull for declaring decorators. It is important in some
cases where method signature has to be preserved between the
original method and the decorated method ; when inspect is used.

We have a feature in rally requiring to preserve signature of the
original method. Here is a link which explains why we need this
dependency: http://fperez.org/py4science/decorators.html

Change-Id: Icef9a043fbf892bb401139902dc38cb6b4bf8030