Switch to SQLAlchemy 0.9
After SQLAlchemy-migrate has been fixed for SQLAlchemy 0.9,
and after the last problem (with Keystone) has been fixed,
we finally can switch to SQLAlchemy 0.9. It is important we
do the switch to SQLAlchemy 0.9.x ASAP, because that is what
distributions (eg: at least Debian) are using. This also
ensure we avoid any future mistakes if we do the gating with
SQLAlchemy 0.9.

This patch sets the cap from <=0.8.99 to <=0.9.99.

Change-Id: I9d2a9e0526359fa1e97df250a63f71496d47367a