Limit psutil to <2.0.0
The psutil package recently released psutil version 2.0.0.  This version
made a significant number of changes that are not backwards-compatible.
We should thus restrict the OpenStack psutil requirement to <2.0.0 until
we can ensure that our code is compatible with the new API and that the
2.0.0 version of psutil has had time to propogate to the various
distros.

Needed by: https://review.openstack.org/#/c/78793/
Change-Id: I63d01235a949cd10ba87d41ad428818daaf9a756