Using happybase >=0.5,!=0.7
The only OpenStack package that uses happybase is Ceilometer, and
it works perfectly well with happybase 0.8. The only thing is that
version 0.7 is borken, 0.5 is too old.

As Debian upgraded to happybase 0.8 (currently in Experimental only,
but it will go in Sid together with the rest of Icehouse) and that
everything worked well with it, the best course of action for
Icehouse is to at least allow 0.8 to be in use.

For Ubuntu that currenly uses the broken 0.7, they may sync from
the version 0.8 currently in Debian experimental (by doing a freeze
exception as well).

Change-Id: I071f9994c55a343766703f0913c65b63bf26627b