request-kerberos
Kerberos requires an additional requests library to perform
SPNEGO

The 0.5 version was just released.  Older Versions of the library
are packaged for distributions already.  They will need to be updated
to this version.  The 0.5 version is required to pass the Python 3.3
requirement enforced on the clients.  Older versions break in py33

Change-Id: I2100915f123c0fea41d5b17d01947901aa0119c5