Add Pillow for stevedore doc build
As part of setting up gate jobs for stevedore, we have turned on
a doc build gate job. That job is failing because it can't install
Pillow. Nothing else uses Pillow, and it is not a runtime dependency
of stevedore.

This change adds Pillow to the stable/icehouse requirements list
so the check-grenade-dsvm job, which tries to verify that the
requirements for the projects are correct across releases,
will allow the new stevedore to use Pillow.

Change-Id: I96c80ce571bae71ef1b867b0c4f60572fa9a4c5c