Add pyscss and django_pyscss for Horizon
PySCSS is a compiler for the SCSS language, an extension of the
Cascading Style Sheets (CSS) language. It is introduced in Horizon
in patch https://review.openstack.org/#/c/90371/

PySCSS is available as python-pyscss in Ubuntu since Saucy:
http://packages.ubuntu.com/search?keywords=python-pyscss&searchon=names&exact=1&suite=all&section=all

and as python-scss in Fedore since 19:
https://admin.fedoraproject.org/pkgdb/package/python-scss/

The django-pyscss needs to be packaged, we have a packager for Fedora
for it.

Change-Id: Iaab685e2a65f682e85b15fc867560580070427f3