Add hacking 0.9.0 series
Hacking 0.9.1 is out (hacking 0.9.0 had a minor bug in it, so its
blacklisted), and we want to start supporting projects to migrate over
to 0.9.x. Expand the upper bound to allow 0.9.x while still
supporting 0.8.x series so we keep the old version in our mirrors as to
not break anything.

Change-Id: I17b7b67135577204de29b82e28526c78d03e8fa5