Add flake8 2.2.2
Flake8 2.2.0 has no new tests but does add parallelization, which is
enabled by default, so we want to move hacking over to use flake8 2.2.2.

The move to flake8 2.2.2 is done in two steps since our
global-requirements system does not allow for non-overlapping changes.

flake8 2.2.0 had a bug with parallelization on windows, 2.2.2 fixed that.

Change-Id: Ib2ccf27cf4ba5cfe2c3f51aebc2859af5a57b998