Disallow alpha version oslo libs in Stable
Recently ran into some problems with cinder.conf file
updates.  Turns out that the current spcification for
oslo.messaging in global-requirements allowed the use
of alpha releases of the messaging library which we
don't want in stable/icehouse.

This patch just changes the requirement to 1.3.0
which should indicate that alpha designated versions
are not acceptable.

Change-Id: Ic4e449ff8654bd58e8594290d814ef0da0f3e8d6