Add libvirt-python for nova.
The python libvirt bindings have long been an implicit dependency for
nova which was expected to be installed via system packages. This was
necessary because libvirt-python could not be installed from pypi. The
libvirt devs have added libvirt-python to pypi which can bind against
libvirt if it isn't ancient as in Precise. We are now testing on Trusty
for master which has new enough system libvirt to install libvirt-python
from pypi.

Add libvirt-python to global requirements so that we can move nova away
from the implicit requirement and onto an explicit requirement.

Change-Id: Ieba2a15eaa1e1b22282179aee8f401607928b18b