Include tempurls in python-swiftclient
python-swiftclient 2.2.0 adds support for tempurls, which is a blocker for
the new (aiming for default) Ironic driver using the Ironic Python Agent.
The Agent needs a temporary URL generated for a Glance image to avoid
passing around admin auth tokens to every machine that is being deployed
to.

Blocked Ironic patch: https://review.openstack.org/#/c/81391/

Change-Id: I35d74f13af9c5cac936bf711e4d2f646531eae60