Don't use SQLAlchemy 9 version < 0.9.7
Versions 0.9.2, 0.9.3 implement Table.kwargs attribute as immutable dict.
It prevents setting defined storage engine for created tables.

Versions 0.9.5 and 0.9.6 are not very good. We could use 0.9.4 though
but not sure if it's worth to explicitly mention it in requirements.

Issue with ummutable dict was discovered during testing of
https://review.openstack.org/#/c/81334/ but issue doesn't occur at the
gate because of used SQLAlchemy version doesn't contain immutable dict.

Change-Id: I2703b2a21365a78fb2e9226680a6ef542b759b4f