Update builds of xstatic-qunit and xstatic-jsencrypt
Both of those xstatic packages had errors in how the libraries
were packaged. This patch updates only the build number of the
xstatic packages, not the versions of the libraries themselves,
so that we can use the fixed builds.

The errors were as follows:
xstatic-qunit -- Missing quinit.css file.
xstatic-jsencrypt -- Packaged just part of the sources,
instead of the merged library.

Those libraries are used only in Horizon at the moment.

Change-Id: I50ae8c2c8d7bad429223d3b3b244333bbb2b1e49