Add urllib3
Glanceclient is using the internal requests.packages.urllib3,
which may lead to all sorts of surprises and errors, as the
urllib3 package is "unvendorized" (ie: removed) from the
requests package in both Debian and Ubuntu. As a consequence,
we want to add urllib3 directly as a package, so we can
replace the "import requests.packages.urllib3" statements by
directly "import urllib3".

Note that version 1.7.1 of urllib3 is available in both Trusty
and Wheezy backports already, so it shouldn't be an issue for
any Debian based installation. Though I don't know for RedHat
or Gentoo based system.

Change-Id: I493fab8667e35bba0835255e347dfdbb27881847