Allow documentation to create sequence diagrams
Sequence diagrams are very useful thing for developer
documentation to include and show (especially seeing that
most of openstack is one big sequence diagram) so it would
be great to include and enable the sequence diagram sphinx
plugin and associated seqdiag software.

It will likely also be required to include:

https://bitbucket.org/blockdiag/seqdiag

Where the docs are being built (so there might be a infra
related changed needed to accomdate this).

Change-Id: I85c163265ea0a10881b5d34f601eea6147d4469d