Using django-nose>=1.2 needed by murano-dashboard
After I packaged Murano, I discovered that its unit tests are
only working with Django-nose >= 1.2. I already wired this
into the build-dependency of the package, but the fact is,
its test-requirements.txt should reflect the reality.

I've checked, and Ubuntu Trusy already has django-nose 1.2.
As for Debian (which I manage), it's not a problem either.
My only concern would be RedHat and Gentoo, for which I
don't know how to check. Please guys, voice your opinion.

Change-Id: I1467842dede515a2dbc98986bf3d3df5f0d13cd3