Cap XStatic-Jasmine version
Jasmine 2.x is causing a Horizon test to time out, blocking the
gate. On master, the test was temporarily disabled because there is
work on-going to add compatibility with 2.x, however this is feature
work that won't be backported to Juno. The library should be capped for
stable releases.

Change-Id: Ibafeb1f7b9fff69d9acc91cb8223d011767a86e6
Related-Bug: #1396954