Bump up taskflow to >=0.6
The newer version of taskflow has functionality that was
extracted from cinder and that can now be shared across
cinder and glance (instead of duplicating code that was
starting to happen). So to avoid further duplication and
technical debt bump up the version so that the needed
removal of duplicated code can occur.

Change-Id: Ia03b2e4888f5f38cc9814b4be3d42c1957232293