Update barbicanclient requirement to 3.x
The 2.x and 3.x branches of python-barbicanclient are API incompatible.
Since the only other option is to allow projects to specify their own
version requirement (which isn't possible), and since Barbican hasn't
actually been promoted yet, it seems reasonable to force an update to
the latest branch of the client now. This way, it will guarantee
everyone is on the new client by the time Barbican is promoted.

Change-Id: Iaf6132d2720351e981f463d9966c22afae8ebdf3