Pin sqlalchemy-migrate to 0.9.1
sqlalchemy-migrate > 0.9.1 breaks unit tests
for nova and > 0.9.2 breaks migrations with
glance when using sqlite (so unit test runs).

The glance migrations are fixed on master
but we should pin sqlalchemy-migrate on the
stable branches if we don't backport the glance
migration fixes from master, and I'm not sure
that backporting changes to migrations should
happen, and they'd still be broken for anyone
before stable/icehouse (which isn't supported
in the community, but is a deployment issue).

Conflicts:
        global-requirements.txt

Change-Id: I39f75fc61f1dfb891c80f7e388dd0b5c201e6a3c
Related-Bug: #1410494
(cherry picked from commit 28c341a37594c2f6419dae7dadc5b75dca862065)