Import cap.py tool to cap  explicit dependencies
This is a forward port of the tool we used to cap stable/juno
requirements. Save a copy of it in master so its available when we are
ready to cut future stable branches.

pin all global-requirements with the version found in pip-freeze
along with a tool to generate the new requirements file

Issues:
* Some versions are only apt-get installable (suds 0.4.1)
* Not all packages are installed in our standard dsvm-tempest env
* Some versions are lower then the minimum requirement we previously had
  because python is awful (libvirt-python)
* Doesn't pin transitive dependencies
* library versions can vanish from pypi and still break us
* Grenade and tempest-full pip-freezes are different and do not work with
each other, so some tweaking is required.

Taking the installed library version from pip-freeze and use that as a
version cap.

Previously used pip-freeze from a tempest-dsvm-neutron-full job.

Change-Id: Iaf48bb069fdd7a19d614ce44b86abd9977c5f0c0