pin oslo.messaging to < 1.6.0
because of eventlet incompatibilities stable/* pins eventlet to <
0.16, however because of fixes and features needed in master, the
current requirements definition requires eventlet >= 0.16.1.

This makes oslo.messaging 1.6.0 not able to work with stable/*
projects.

The other alternative would be to make oslo.messaging not specify
eventlet version, and only take it from the projects. However that's a
larger fix to unwind.

... ALSO ...

Pin barbicanclient because it creates a requirements wedge with juno
cinder.

Change-Id: I7a0be6d1b20a6d4220600dd58ff0c0ecc068d0da