Add setuptools to global requirements
Some things (namely stevedore) already depend on setuptools
but they rely on the fact that setuptools is often implicitly
available by it's nature of being a commonly installed tool.
However setuptools is not a mandatory library, and pip in
particular can install from Wheels without setuptools being
present. To rectify this setuptools needs added to the
requirements of any project that depends on it at runtime.

Change-Id: Icd83a9aed56942108f240824231216b05b873b3b