Do not break on projects without setup.cfg
setup.cfg is used by pbr instead of setup.py.
Projects not using pbr while not have setup.cfg at all
or will have it with completely different contents.

There's already a check in update.py whether or not
the target project is using pbr, but it relies on
presence of setup.cfg, which may not be the case.

After this patch update.py will try to read setup.cfg
only after it detected that pbr is used.

Change-Id: I21fbbe159007b25db546fa6d36721e018d1dca24