Add toposort test requirement
This is a test requirement for Nova change
Ic4dee3126dda2a25455235fea6a6574b64cd51a3 .

toposort implements a topological sort in Python. The above change
uses it to ensure correct ordering wrt referential integrity
constraints when reconstructing a MySQL schema from a cache during
unit tests.

The library is fairly simple. I would expect it to be updated rarely.

It is python 3 compatible.

It uses the Apache 2.0 licence.

It is not packaged for either Ubuntu or Fedora as of February 2015.

Its function is not covered by any other currently required library
that I'm aware of.

A feasible, but undesirable, alternative to using this package would
be to implement a topological sort function in OpenStack.

Change-Id: I7c9c88088228514e37649a8d06d94c43509962a5