Do not sync requirements to possibly capped library's master branch
We've hit a situation where a project (pycadf) is listed and capped
in stable/icehouse requirements.txt.  pycadf is also listed in
projects.txt. As part of the requirements integration test, we are
attempting to update pycadf's master branch with stable/icehouse
requirements, and failing because of new dependencies in pycadf that
are not managed by stable/icehouse's GR.

This updates integration.sh so that when we're testing a stable branch,
we are only syncing requirements to other projects have stable branches.
This avoids trying to sync stable requirements into library's master
branches.

Change-Id: I3b8845ad8e013cf6747995df5a62cc4f2ee85e01