Bump oslo.log for critical nova regression
Once Nova switched over to oslo.log, we realized that "instance"
objects were printed wrong in the nova logs. Essentially instead
of just the uuid, the entire __repr__ of the instance object was
printed in the logs, making them unreadable. This was fixed in
I2d1b09015bead2012327f973f45f3f6728160c69 and oslo.log 0.4.0
was released.

Related-Bug: #1426544
Change-Id: I982b165ed64494a14dcd6b2e78d3f43a3f75aa53