Add contextlib2 lib to aid in py2/py3 compatability
The contextlib2 provides functionality of that exists
in the newer contextlib of python 3.2 and newer for
projects that wish to work seamlessly in 2.6/2.7 and
3.2 and newer they need this package to be able to
seamlessly use those python versions functionality.

This package exists on at least ubuntu 14.04 and
therefore is already being used/supported (and it
seems to have a fedora package as well).

TaskFlow would like to use it to provide use an
equivalent of the contextlib.ExitStack functionality
that exists in this library (and natively exists in
3.2 and newer). Oslo.db should also be able to take
advantage of the same functionality to replace some
test code (and likely other use-cases exist as well).

Change-Id: I328a29d0f095204dfbe175a9044e26ed653cb01d