Bump keystonemiddleware requirement
Keystonemiddleware 1.5.0 was released 2015-03-11.

We are continuously adding new features to auth_token middleware,
however on the service side we continually check for environment
variables to discover what version of middleware is installed. Bumping
this globally should allow us to remove some older code paths and rely
on the presence of newer functionality.

Whilst not strictly required, bumping this now ensures that at least
this new version of keystone middleware will be present in kilo
distributions.

Change-Id: I45275f60e2fa1f0a372b2a1f82f43640c73f3b6f