Capped more global-requirements entries
Now used ceilometer pip-freeze output generated from py27 tox env.

The only difference between the autogenerated version of the file and
the one included in the patch is pymongo cap, which is lowered to avoid
pymongo 3.0+ that broke stable branches in addition to master.

Added pip freeze file for ceilometer unit tests for reference.

Related-Bug: #1441393
Change-Id: I742842c67576417a46eea52f10a74abd4d5b6656