create a separate section for pinned requirements
Pinning a requirement is a very extreme measure, so we should be
really explicit about the kinds of things we allow pins for, and
why. Moving a pinned requirement is also incredibly disruptive, so
this attempts to set a policy about when they should be updated as
well.

All the linter like things are moved to the end of the file and called
out with this policy inline.

Change-Id: Ia91bb7adc25e845d993811f85e59334f31ff9f47