Remove tempest from projects.txt on stable
Because, of branchless tempest the requirements file for tempest will
only have a master branch available. The requirements in that will
likely conflict with the global requirements list on the stable
branches. So this commit removes tempest from being synced with
global requirements on stable branches which is needed to avoid
conflicts on dsvm jobs where devstack forces a sync to the stable
version of global requirements.

Change-Id: I2ac2923b35a97a6f7fe158882bf7c4c92f0a9277
(cherry picked from commit 75729856af0b250dbbf8c9f0ab77ad0489f8d208)