Raise cap on tooz to <0.13
tooz does not have a stable/juno branch yet and tooz 0.12 has uncapped
requirements on zake and kazoo which is breaking with zake 0.2.2 which
blocks kazoo 2.1 - which is what gets pulled in.

So we're going to create a stable/juno branch for tooz from the 0.12
tag, sync it's requirements with g-r and then release tooz 0.12.1 so we
need to raise the cap here to allow for 0.12.1.

Once this is in, it'll sync to ceilometer on stable/juno and we'll need
to merge that also to fix stable/juno and stable/kilo changes.

Partial-Bug: #1459322

Change-Id: I4e720c734f51b9e8375108b00d3f3a91ede3c77e