Enable 'python_version=='2.6' for futures and enum34
Enable futures for 2.6 this until the discussion around oslo.messaging
need for this (and its support for 2.6) and its planned usage of the futurist
library for the liberty release.

Also some of the current oslo libraries (taskflow) still support 2.6
for the short-term (until a few other reviews merge) so we need to enable
its usage there still for enum34.

This is also breaking the futurist, automaton, gate and the taskflow check
gates as those are still on 2.6 (this may change soon).

Soon the following will make that less relevant (but these haven't merged
yet):

- https://review.openstack.org/#/c/193349/
- https://review.openstack.org/#/c/193354/
- https://review.openstack.org/#/c/194860/

Change-Id: I9256563a6c2cc455043f0d61cd27b033d37302ef