Cap fixtures<1.3.0
More recent versions of fixtures introduce behavioral changes that,
while not strictly API breaks, do break the tests for some projects where
fixtures are cleaned up more than once. We could backport the fixes we
have made in master, but that would introduce an excessive amount of
churn in the stable branch.

Change-Id: I0b2a43b88cff367684c4d2918305de1547f4411c